Fix adding the current guest participant again and again
"fetchSingleConversation" is periodically called in the guest UI to
update the conversation data, so a new participant for the current guest
was added every time that the conversation was added (unlike
participants adding again the same conversation is not a problem, as the
new data replaces the previous data).

As the current participant is automatically added (if needed) to a
conversation when that conversation is added there is no need to add the
participant from "fetchSingleConversation".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>