Fix adding participant with previous session for current guest
In order to fetch a public conversation a guest has to join it first;
otherwise just a set of limited data about the conversation is received.

Moreover, if a guest page is reloaded and the conversation is fetched
before joining it again the previous session will still be active, so
the full conversation data will be received, including the previous
session, which will be automatically added to the participants store
associated to the current guest. However, once the guest joins again a
new session will be generated, so the participants store ends with two
participants for the current guest, each one with a different session.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>