Add a "constrained layout" for call view to be used in the sidebar
Currently this mimics the layout used in Talk 7 when the call view was
shown in the Files app sidebar (smaller avatars, videos and paddings, no
screensharing button...), but even once the call view is moved to the
new layout a special handling will be needed when the space is limited.

Ideally the call view should automatically change between layouts based
on its size (probably using something like
ResizeObserver/MutationObserver), but for the time being it is
explicitly enabled when the call view is shown in the Files app sidebar.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>