Enable HTML markup in contenteditable div
By default the "vue-contenteditable-directive" sets the text of a
contenteditable div using the "innerText" property, which causes any
HTML markup to become escaped plain text inside the div. In order to
provide rich text messages with mentions real HTML elements are needed,
so the "dangerousHTML" option needs to be enabled so the text is set to
the "innerHTML" property instead.

When real HTML elements are used the browser adds new lines using "<br>"
and spaces that would otherwise be collapsed into one (like several
white spaces in a row) using "&nbsp;". Therefore now the raw text of the
contenteditable div needs to be converted to the format expected by the
server (that is, "\n" and regular white spaces).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>