Add custom template for elements in the autocompletion panel
In order to show the user label in the autocompletion panel but set the
user id in the content editable both attributes need to be included for
every vue-at member. When the members are objects instead of a string
the "name-key" property needs to be used so vue-at knows which attribute
to use to filter the panel based on the text after the "@". However, the
default embedded item uses that attribute, so a custom template is
needed to set the user id in the content editable. In turn, this causes
the content editable contents to be set as HTML elements instead of
plain text, so the HTML element of the mentions need to be converted to
plain "@userId" before sending them to the server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>