Replace default vue-at style sheet with a wrapped version
The wrapped version roots all the rules of the default style sheet on
".talk.candidate-mentions", so the rules affect only the candidate
mentions of Talk. Otherwise, as the default rules are too broad (as they
use just ".atwho-XXX") they could mess with the autocompletion of other
elements (like the comments tab in the Files app).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>