Fix autocompletion of mentions in public share auth page
The autocompletion panel is reparented to the body element and placed
using an absolute position. However, as the body element used a flex
layout the absolute position of the panel was ignored. Therefore now all
the body elements are wrapped in another div element, which becomes the
flex element and acts like the body element, so the real body element is
no longer flex and reparenting the autocompletion panel to it then works
as expected.

Besides that the CSS for the autocompletion panel were not included in
public share auth pages.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>