Add support for opening files through Viewer
Note that the "path" property was already being passed to FilePreview,
although it was not explicitly declared.

At this time the Viewer can open only internal files, which are
available only for registered users. Therefore the Viewer is not loaded
for guests.

Also, for the time being, due to a limitation in the Viewer API, when a
file is opened the Viewer allows the user to iterate through all the
files in the directory of that file instead of through all the files in
the conversation, as it would have been expected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>