Add support for muting other participant from the call view
This should probably be done adding another entry to the menu of each
participant in the participant list instead, but it is currently not
possible to get the signaling session id from there. Until that is
solved doing it from the call view is a simple way to add support for
this.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>