Limit being able to force mute others only to moderators
Although the signaling message could be filtered by the signaling server
that would require changes in both the internal and the external
signaling server. For simplicity the clients are the ones that check if
the received message comes from a moderator in order to honour it or
not.

Note that, when the external signaling server is used, it is currently
not possible to know which guest is certain peer. Moreover, no matter
which signaling server is used, guests can not get the full participant
list, so they can not know if a guest is a regular guest or a guest
moderator. Due to all this guest moderators can not force mute other
participants; only logged in moderators can.

As "forceMute" messages are taken into account only if they come from a
moderator now the button to force mute another participant is hidden for
non-moderators.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>