Send "forceMute" message to all the participants
Now the "forceMute" message includes the ID of the peer to be muted. If
the participant receiving the message is the one being force muted it
will mute itself, but other participants receiving the message will
locally mute the specified participant.

This ensures that even if a client being force muted does not honour the
message from the point of view of the other participants it will be
muted anyway (although only for those participants already in the call;
participants joining later would need to force mute themselves that
rogue client).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>