Move audio element from view to model
If a Video view is removed its audio element no longer plays the audio.
This can happen, for example, when closing the sidebar in the Files app,
which removes the call view and with it its Video views. But in that
case, even if there is no call view, the call is still active. To ensure
that the audio for another participant is played even if there is no
view for it the audio element must be part of the model instead.

Note that when closing the sidebar the audio was still heard, but that
is caused by a duplicated audio element created in the SimpleWebRTC
code (which will be fixed in a following commit).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>