Do not create audio and video elements without container in SimpleWebRTC
SimpleWebRTC automatically created an audio and video elements when a
peer stream was added. Those elements were ignored by the Talk code
(which creates its own elements as needed), so they were not needed but
active nevertheless.

In the case of the audio element this caused its audio to always be
played, which would keep the audio from the other participant active
even if the audio element handled by Talk was removed or muted.

For consistency with the rest of its code now SimpleWebRTC creates the
audio and video elements for a peer only if a container for them is
provided.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>