Update participants store from conversation data for guests
Guests can not fetch the full participant list, so until now the
participants store was only populated with the guest itself. However,
the conversation data also includes a list of the users (but no guests)
in the conversation, so the participants store can be populated with
that data. This will make possible to know if a message is sent by a
moderator or not.

Note that when the external signaling server is used guests can also
know which other guests are in the conversation from the participants
update signaling message. However, that data is not used to populate the
participants store because the conversation data is enough to know the
moderators and to keep a consistent behaviour when using the internal
and the external signaling servers.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>