Filter control messages not sent by moderators
The internal signaling server now filters signaling messages with type
"control" that are not sent by logged in moderators (including the
owner). This will be used, for example, to filter force mute messages
directly in the server, without needing to rely on the clients filtering
them.

The external signaling server already supports control messages,
although with a slightly different format (instead of being a subtype of
"message" messages they are their own "control" message). Therefore,
internally the WebUI uses the same format as the internal signaling
server for control messages and formats them as needed when sending them
to or receiving them from the external signaling server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>