Send participant permissions to the external signaling server
In order to filter control messages the external signaling server needs
to know the permissions of each participant. Those permissions need to
be set when the participant joins the room or if the participant type is
changed.

All participants should have permissions to publish media and screen,
but only logged in moderators should have permission to send control
messages.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>