Honour media constraints when joining a call
The participant call flags are now honoured as media constraints when
joining a call. Therefore, now only the media specified in the flags is
allowed to be sent (although whether it is sent or not depends on the
available media).

Note that the default constraints now need to be hardcoded in
"LocalMedia.start". Otherwise if no constraints are given the object
that contains the default ones could get modified depending on the
available media (for example, if there is no camera connected), and the
updated ones will be used from that moment on even if the media has
become available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>