Fix attachment folder in Talk sidebars
The attachment folder was set when the left sidebar was loaded. That
component is loaded only in the main Talk UI, so the attachment folder
provided in the initial state was ignored in the Talk sidebars. Now the
initial state for the attachment folder is used by default in the store,
so it applies everywhere.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>