Add helper to throttle video quality based on the call state
In order to (try to) reduce the overall resources needed, but specially
the load on the CPU of the clients, now the quality of the sent video is
automatically adjusted based on the number of participants with audio
and video in a call.

There is a threshold for each quality. When the number of participants
(besides the self participant) with audio or video in the call is above
or equal the threshold that quality is used. In case of clash between
the thresholds the most restrictive one wins. The default thresholds are
basically guesses; whether a different value would be more suitable is
something that is still to be seen :-)

Both the audio and video thresholds refer to enabled audio and video.
That is, if a participant has audio or video available but has muted
himself and disabled the video then it will not be taken into account in
the thresholds (but if you hide the video of someone else only for you
that participant will still be taken into account).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>