Set also minimum values for the resolution in the constraints
When changing from a constraint with smaller max values to one with
larger max values applying the new constraint may succeed, yet the video
stream quality may not have changed, as the previous state still matches
the new constraints. In those cases the resolution can be increased by
providing too min values.

Additionally, when max and min values are provided to increase the
resolution some browsers may settle for the min value, even if it is
possible to go higher. Therefore an "ideal" attribute is provided too
now.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>