Prevent "No bytes received" to be logged by default
As the catch did not have any function the caught error was logged by
default by the browser. However, when a connection is established with
the HPB it is perfectly normal that no bytes are received at first, so
there is no point in logging that. There are some corner cases in which
a connection may be established but no bytes are ever received, but in
that case it can be detected by other means and the messages just flood
the log providing no value and making more difficult to debug the issue.
Due to all that now the "No bytes received" messages are no longer shown
in the logs.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>