Fix microphone not available when camera can not be accessed
"getUserMedia" is first called for both audio and video and then, if the
camera was not found, it was called again only for audio. In some cases
(for example, in Windows if the camera device is being used by another
application, or if permissions for the camera are not granted at a
system level) it could happen that a camera device is found but it can
not be accessed by the browser. If that happened "getUserMedia" was not
called again only for audio and the microphone was also seen as not
available, even if the problem was only with the camera.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>