Fix guest name sent when using the internal signaling server
When the external signaling server is used the guest name is continuosly
sent through the data channels. However, when the internal signaling
server is used the guest name is expected to be sent only when it is
modified.

This is not strictly needed for the web, but it might be expected by the
mobile apps, and it makes the behaviour of the participant models more
consistent between the internal and external signaling server (as
otherwise the "name" attribute would not be set when using the internal
signaling server).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>