Fall back to the store name only if it is not set in the model
When the external signaling server is used it is not possible to get a
guest name from the store, as the signaling session ID and the Nextcloud
session ID are different and can not be linked. Due to this, if the
model name is unconditionally overriden with the store name then guest
users are always shown as "Guest". Therefore, now getting the name from
the store is used only when the participant model does not have a name
(which basically just show the "Guests" placeholder, as in other cases
the name would already come from the model).

Note that if the participant name is undefined it is a special case in
which no name is shown for consistency with the behaviour with
registered users (which is not shown until the participant has been
connected). Unfortunately this implies that if a participant does not
have audio nor video then no connection is made and her name is not
shown; this will be fixed in the future once it is possible to link
signaling session IDs and Nextcloud session IDs.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>