Ignore drag and drop events if files are not involved
If a drag and drop events does not involve files the drop hint should
not be shown, and the drop event does not need to be handled either. The
drag event may not contain yet the files in its "dataTransfer.files"
property, but at least one of the types included in its
"dataTransfer.types" property will be "Files".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>