Add documentation about testing the TURN server
In order to know if a TURN server is working it is useful to actually
test the data transfer through the TURN server instead of just checking
if it generates relay candidates.

When an RTCPeerConnection is created an RTCConfiguration can be provided
to configure it. One of the values it accepts is "iceTransportPolicy",
which can be either set to "all" to gather all types of candidates or to
"relay" to only gather relay candidates.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>