Do not fetch all conversations on room participants update events
The room participants update events are sent by the external signaling
server:
  -When a user resumes the session
  -When an internal client joins or leaves the conversation
  -When the BackendNotifier sends a "participants" event (that is, when
the type of a participant changes, when the name of a guest changes or
when guests are cleaned)
  -When the BackendNotifier sends a "incall" event (that is, when a
participant joins or leaves the call)

In all those cases it is not really needed to fetch all the
conversations to get the updated state, but only the current
conversation. Note that it is not enough to fetch just the participants,
as for example when a call is started that modifies some attributes in
the conversation, like "hasCall".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>