Fix "joinCall" of standalone signaling not returning a promise
"Signaling.Standalone.joinCall" should return a promise, like done for
"Signaling.Base.joinCall".

The standalone signaling object is able to defer joining a call if
called before its room has been joined. However, it only takes into
account the last room that a call was tried to be joined in. This is
still the case now when the promise is returned; if "joinCall" is called
on a different room while a previous was pending the previous one is
just rejected (although if it is in the same room the previous promise
is reused).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>