Do not fetch all conversations when joining or leaving a call
When the current user joins or leaves a call there is no need to fetch
all the conversations, but just the current conversation and the
participants. As when any user, including the current one, joins or
leaves a call a room participants update event is sent by the external
signaling server the current conversation and the participants are
already gotten, so it is not needed to add an explicit handling for
that.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>