Fix video marked as available due to screen sharing
When the external signaling server is used the incoming audio and video
stream is checked to know if there is actually audio and video being
received. However, that check was performed also for the screen peer. As
the audio/video peer and the screen peer of the same participant have
the same id that caused the video to be marked as available even if the
other participant was not sending any video. Now the check is performed
only on the audio/video peer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>