Fix unsharing screen of other peers when one screen was unshared
When the "unshareScreen" signaling message is received an internal
"unshareScreen" with the id of the peer that sent the message is
emitted. However, when the internal "unshareScreen" was handled the id
of the other peer was expected to be in the "from" property instead of
the "id" property. As the "id" property was undefined this caused that
all the screen peers were processed and ended instead of only the one
that sent the message.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>