Fix grid style when the empty call view is shown
The empty call view is shown in the grid along with the local video if
there are no other participants in the call. In that case the rows and
columns of the grid will be 0, so they need to be overriden to show the
empty call view above the local video with both having the same height.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>