Fix file attributes passed to Viewer
When an explicit list of files is provided to the Viewer the file info
of the file is not fetched by the Viewer and the given file attributes
are the ones used. Due to this all the currently known file attributes
that are also used by the Viewer should be provided.

Note that the given mime type is overriden by the Viewer when only a
single file is provided, but it was included anyway for consistency.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>