Show sidebar when viewer is opened
The Viewer expects that, if the app has a sidebar, an
"OCA.Files.Sidebar" object exists. This object should contain a
"state.file" property with the name of the file currently shown (or an
empty string if the sidebar is closed), and "open" and "close" functions
to be called by the Viewer when opening and closing the sidebar is
triggered by the Viewer.

The sidebar adjusts its width to the width of the sidebar once the
opening ends. The sidebar comes from nextcloud/vue, and its opening is
animated with a transition, so "OCA.Files.Sidebar.open" needs to wait
until the transition ends to be resolved.

Besides the "OCA.Files.Sidebar" object the Viewer also expects that the
sidebar element in the DOM has an "app-sidebar" id. In nextcloud/vue >=
2.0.0 the id of the sidebar was changed to "app-sidebar-vue", so it
needs to be overridden to be found by the Viewer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>