Always try to use microphone if camera failed
When getting the local media it is first tried to get both microphone
and camera and, if that fails, then it is tried to get only the
microphone.

However, this was done only when certain errors happened while trying to
get both the microphone and the camera. Now, no matter the error, it is
tried too to get only the microphone; if the media is still unavailable
it will fail again, but this should help with some further scenarios not
covered yet in the previously handled errors in which camera is not
available but microphone is.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>