Handle error when joining room in the external signaling server
Joining a room in the external signaling server can fail (for example,
if the Nextcloud server took too long to send a response and joining
timed out). When that happens an error is returned, but the error was
ignored. Now, when joining a room in the external signaling server fails
it is tried again and again until the room is joined.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>