Prevent infinite loop in datachanel open re-transmission
As there could be two data channels with the same name (created by the
remote and local side) and they will be connected at different time, it
is possible to get an onopen event while the other connection is still
in connecting state.

When the above situation happens, pendingMessages cannot be resent
because the check is on the remote-to-local connection, which is still
in "connecting" state, while the onopen is called for local-to-remote.

At this moment the sendDirectly function will queue the message again
and the loop will try to re-transmit the message again on the next
iteration where it will be queue again, and so on.

To prevent this infinite loop, a copy of the pendingMessages queue is
created and resend happens on this "new" list. This way in case the
channel is already opened the transmission will happen, and in case
the channel is not yet opened it will be re-queued for the next onopen
when the other connection is established.

Signed-off-by: Vladislav Hristov <vvtthristov@gmail.com>