Add cache for video constraints
When video constraints are adjusted based on the quality applying the
default constraints may fail, and a relaxed but roughly equivalent set
of constraints has to be found to be applied instead.

Until now the check for the valid constraints was done every time the
quality was changed. Now a cache is added to keep track of valid
constraints for each quality to apply them directly.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>