Take round trip time into account when calculating the quality
A high round trip time means that there is a delay in the connection,
which is in itself bad (specially in a conversation instead of in a
"monologue").

Besides that a delay in the connection can also cause that some packets,
even if they are not lost, are discarded anyway to try to keep the
playing rate in real time.

The round trip time is only available for sent data, but not received
data.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>