Show warning on the button that could solve the issue
Before the warning tooltip about a bad connection quality was shown on
the button for the media that had the issue. Now the tooltip is shown
instead on the button for the media that could solve the issue. So, for
example, "Audio is bad, disable your video" before was shown on the
audio button, but now it is shown in the video button.

If only one type of media is being transmitted then the tooltip is shown
on the button for that media ("so "Audio is bad" is shown in the audio
button, as it is just an information but there is nothing that can be
done about it).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>