Do not reuse media constraints between independent media requests
When getting the user media for the first time both audio and video are
requested. If video is not available then the user media is got again
requesting only audio. However, as the media constraints were reused in
any following user media request as soon as video failed once it was
never requested again. Therefore if a camera was plugged in later it was
ignored until Talk was restarted.

Now getting the user media always tries first with both audio and video
and then fallbacks to trying only with audio, even if video was not
available in a previous request.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>