Add components to select and preview media devices
MediaDevicesSelector is a dropdown to select a media device of a
specific kind.

MediaDevicesPreview has selectors for audio and video inputs that modify
the devices to be used in MediaDevicesManager. Additionally it also
shows a preview of the selected devices with a volume bar in the case of
audio or a video element in the case of video.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>