Fix speaker promotion with newer Janus versions
When Janus is used even if it is possible to open several data channels
with different labels and send data on them all the messages are
received in the first data channel opened.

Although currently several data channels are opened in Talk in practice
only the "status" data channel is used (and messages received in data
channels with a different label are ignored). This is also the first
data channel opened when Janus is used, so until now everything happened
to work (with Janus 0.4.3).

However, with newer Janus versions the data channel messages are
received instead on a data channel opened by Janus, which is named
"JanusDataChannel". It is not possible to change that behaviour on Janus
except by patching it, so for broader compatibility the messages
received in "JanusDataChannel" are treated like messages received in the
"status" data channel (this was partially done already; this commit
finishes the change).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>