Do not update streams again while waiting for a previous one
When an audio or video device is selected a stream from that device is
requested to show its preview. The stream is resolved asynchronously, so
while that happens the user could select a different device, which in
turn will request another stream. As only a single stream for each type
is expected to be active in the previews this could lead to streams
being opened but never closed.

Now this is enforced by preventing further stream requests while a
previous one has not been completed yet. If several stream updates are
triggered while waiting for a previous one once that previous one is
finished a new stream will be requested for the last selected device.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>