Update selected devices from the getUserMedia results
In Firefox the dialog to grant media permissions allows the user to
change the device to use. In that case the device selected in the dialog
will be the one used, no matter which one was originally requested. Now
this is taken into account and the selected device is updated to reflect
the one selected by the user in the dialog.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>