Handle empty "deviceId" when media permissions are not granted
In latest MediaDevices spec if permanent media permissions have not been
granted and there is no active stream "enumerateDevices" returns at most
one device of each kind, and all of them with empty attributes
(including the deviceId) except for the kind.

This is already partially implemented by Chromium, so devices with an
empty "deviceId" need to be taken into account.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>